requirements.txt

flask
transformers
PyPDF2
python-docx
matplotlib